attachments split out, but it's pretty buggy
